requirements.txt
# Django Core
Django>=4.2

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database (SQLite by default, PostgreSQL optional)
psycopg2-binary>=2.9.9

# Geolocation (for nearby events)
geopy>=2.4.1

# Email and utility helpers
requests>=2.31.0
django-environ>=0.11.2
python-dotenv>=1.0.1

# CORS headers for frontend connection
django-cors-headers>=4.4.0

# Image uploads and file handling
Pillow>=10.2.0

# Server / Deployment
gunicorn>=21.2.0

# Optional tools (for admin / logging / debugging)
whitenoise>=6.6.0
